# perf: use total cmp for ordered float