TL;DR: `PartialOrd` impl for `f32` requires two cmp instructions and `PartialOrd::lt` is not reduced to a single instruction. With `total_cmp` we avoid this overhead. See https://godbolt.org/z/ez4EaEs3W

Note: maybe we could gain more perf by implementing a `NaN-less` cmp?

V1:
```
Running query with nprobes: 1, refine_factor: 0
Time: 652.877µs

Running query with nprobes: 10, refine_factor: 0
Time: 818.418µs

Running query with nprobes: 50, refine_factor: 0
Time: 1.406433ms

Running query with nprobes: 100, refine_factor: 0
Time: 1.747243ms

Running query with nprobes: 512, refine_factor: 0
Time: 5.754752ms
```

V3 before perf fix
```
Running query with nprobes: 1, refine_factor: 0
Time: 752.659µs

Running query with nprobes: 10, refine_factor: 0
Time: 2.379638ms

Running query with nprobes: 50, refine_factor: 0
Time: 9.518057ms

Running query with nprobes: 100, refine_factor: 0
Time: 17.623498ms

Running query with nprobes: 512, refine_factor: 0
Time: 39.43036ms
```

V3 after perf fix
```
Running query with nprobes: 1, refine_factor: 0
Time: 644.933µs

Running query with nprobes: 10, refine_factor: 0
Time: 1.512715ms

Running query with nprobes: 50, refine_factor: 0
Time: 5.974101ms

Running query with nprobes: 100, refine_factor: 0
Time: 9.848213ms

Running query with nprobes: 512, refine_factor: 0
Time: 23.058848ms
```